fix FingerprintService and MatchingService in api module

+++ apps/api/src/main/java/com/scalebiometrics/api/service/MatchingService.java
@@ -26,7 +26,7 @@ public class MatchingService {
                 .setTraceId(traceId)
                 .setProbeRid(requestDto.getProbeRid())
                 .setTopK(requestDto.getTopK())
-                .setThreshold(requestDto.getThreshold())
+                .setThreshold((int) requestDto.getThreshold())
                 .setTimeoutMs(2000); // Default timeout
 
         if (requestDto.getProbeTemplate() != null) {
